+++ docs/menu.js
@@ -23,16 +23,16 @@ export default {
                 {
                     category: 'Form controls',
                     pages: [
-                        { name: 'Autocomplete', path: '/documentation/autocomplete' },
+                        { name: 'Autocomplete', path: '/documentation/autocomplete', isUpdated: true },
                         { name: 'Checkbox', path: '/documentation/checkbox' },
                         { name: 'Datepicker', path: '/documentation/datepicker', isUpdated: true },
-                        { name: 'Field', path: '/documentation/field', isUpdated: true },
-                        { name: 'Input', path: '/documentation/input' },
+                        { name: 'Field', path: '/documentation/field' },
+                        { name: 'Input', path: '/documentation/input', isUpdated: true },
                         { name: 'Radio', path: '/documentation/radio' },
-                        { name: 'Select', path: '/documentation/select' },
+                        { name: 'Select', path: '/documentation/select', isUpdated: true },
                         { name: 'Switch', path: '/documentation/switch' },
-                        { name: 'Taginput', path: '/documentation/taginput', isNew: true },
-                        { name: 'Timepicker', path: '/documentation/timepicker', isNew: true },
+                        { name: 'Taginput', path: '/documentation/taginput', isUpdated: true },
+                        { name: 'Timepicker', path: '/documentation/timepicker', isUpdated: true },
                         { name: 'Upload', path: '/documentation/upload' }
                     ]
                 },
@@ -41,12 +41,12 @@ export default {
                 { name: 'Message', path: '/documentation/message' },
                 { name: 'Modal', path: '/documentation/modal' },
                 { name: 'Notification', path: '/documentation/notification' },
-                { name: 'Pagination', path: '/documentation/pagination' },
-                { name: 'Snackbar', path: '/documentation/snackbar', isUpdated: true },
-                { name: 'Table', path: '/documentation/table' },
-                { name: 'Tabs', path: '/documentation/tabs' },
+                { name: 'Pagination', path: '/documentation/pagination', isUpdated: true },
+                { name: 'Snackbar', path: '/documentation/snackbar' },
+                { name: 'Table', path: '/documentation/table', isUpdated: true },
+                { name: 'Tabs', path: '/documentation/tabs', isUpdated: true },
                 { name: 'Tag', path: '/documentation/tag' },
-                { name: 'Toast', path: '/documentation/toast', isUpdated: true },
+                { name: 'Toast', path: '/documentation/toast' },
                 { name: 'Tooltip', path: '/documentation/tooltip' }
             ]
         }