+++ docs/pages/components/autocomplete/api/autocomplete.js
@@ -23,11 +23,11 @@ export default [
                 default: '<code>value</code>'
             },
             {
-                name: '<code>max-results</code>',
-                description: 'Total of results/options to show',
-                type: 'Number',
+                name: '<code>open-on-focus</code>',
+                description: 'Open dropdown list on focus',
+                type: 'Boolean',
                 values: '—',
-                default: '<code>6</code>'
+                default: '<code>false</code>'
             },
             {
                 name: '<code>keep-first</code>',